Cleaning the HomeViewController
