+++ project-saarah/controllers/HomeViewController.swift
@@ -9,46 +9,8 @@
 import UIKit
 
 class HomeViewController: UIViewController {
-
-    var tableView = UITableView()
-
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = AppStyleGuide.Colors.background.uiColor
-        configureTableView()
-    }
-
-    func configureTableView() {
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        tableView.separatorStyle = .none
-        view.addSubview(tableView)
-
-        NSLayoutConstraint.activate([
-            tableView.topAnchor.constraint(equalTo: view.topAnchor),
-            tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            tableView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
-        ])
-
-        tableView.delegate = self
-        tableView.dataSource = self
-    }
-
-}
-
-extension HomeViewController: UITableViewDelegate, UITableViewDataSource {
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        let headerView = EmptySectionHeaderView()
-        return headerView
-    }
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 5
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = SelectableCell()
-        cell.label.text = "Item"
-        return cell
     }
 }
